Start moving these calls into a new class.
