+++ Runtime/TechTalk.SpecFlow.csproj
@@ -60,6 +60,7 @@
     <Compile Include="Assist\SetComparer.cs" />
     <Compile Include="Assist\TableDifferenceResults.cs" />
     <Compile Include="Assist\TableDiffExceptionBuilder.cs" />
+    <Compile Include="Assist\TableStuff.cs" />
     <Compile Include="Assist\TEHelpers.cs" />
     <Compile Include="Assist\InstanceComparisonExtensionMethods.cs" />
     <Compile Include="Assist\IValueComparer.cs" />
